server/pyproject.toml
[project]
name = "languagelearn-server"
version = "0.1.0"
description = "LanguageLearn API server"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115",
    "psycopg[binary,pool]>=3.2",
    "fsrs>=6.0",
    "webauthn>=2.0",
    "pyjwt[crypto]>=2.9",
    "bcrypt>=4.2",
]

[dependency-groups]
dev = [
    "ruff>=0.9",
    "pyright>=1.1",
    "pytest>=8.0",
    "pytest-cov>=6.0",
    "pytest-asyncio>=0.25",
    "httpx>=0.28",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "D",       # docstrings — not required
    "ANN",     # type annotations — pyright handles this
    "COM812",  # trailing comma — conflicts with formatter
    "ISC001",  # single-line implicit string concat — conflicts with formatter
    "TC002",   # type-checking block — breaks runtime for FastAPI Depends() annotations
    "TC003",   # type-checking block — breaks runtime for FastAPI Depends() annotations
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101", "S105", "S106", "PLR2004", "PLC0415", "ERA001", "PT018"]
"db/queries/**" = ["PLR0913", "S101"]   # query functions map to table columns; assert for INSERT RETURNING
"auth/**" = ["PLR0913"]                  # auth functions have many config params
"routes/**" = ["PLR0913"]                # route handlers have many dependencies

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "--cov=. --cov-report=term-missing --cov-fail-under=80"

[tool.coverage.run]
omit = ["tests/*", ".venv/*"]

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "strict"
venvPath = "."
venv = ".venv"
reportUnknownArgumentType = "warning"
reportUnknownParameterType = "warning"
reportMissingTypeArgument = "warning"
